config/components: rename islandMesh.ts back to island.ts for consistency with other component configs (sky, gem, bushes, etc.)
