+++ web/src/scene/components/island/islandShader.ts
@@ -6,7 +6,7 @@
 // day/night cycle.
 
 import * as THREE from 'three';
-import { ISLAND_MATERIALS } from '@/config/components/islandMesh.js';
+import { ISLAND_MATERIALS } from '@/config/components/island.js';
 
 const vertSrc = /* glsl */ `
 attribute vec3 color;
