+++ web/src/scene/components/trees/treePlacement.ts
@@ -27,7 +27,7 @@ import * as THREE from 'three';
 import { TREES } from '@/config/components/trees.js';
 import { FOOTPRINT } from '@/config/components/footprint.js';
 import { BUILDING_DIMENSIONS } from '@/config/components/buildings.js';
-import { ISLAND_GEOMETRY } from '@/config/components/islandMesh.js';
+import { ISLAND_GEOMETRY } from '@/config/components/island.js';
 import { getWorldBounds } from '../../layout/worldBounds.js';
 import {
   buildTopPolygon,
@@ -36,7 +36,7 @@ import {
 import { islandSeedFromBounds } from '../island/islandMesh.js';
 import { StreetAxis } from '@/types';
 import type { Building, BuildingPath, CityBbox, CityLayout, Street } from '@/types';
-import type { IslandGeometryConfig } from '@/config/components/islandMesh.js';
+import type { IslandGeometryConfig } from '@/config/components/island.js';
 
 interface Rect {
   minX: number;
