+++ web/src/scene/components/trees/treePlacementClient.ts
@@ -10,7 +10,7 @@ import { placeTrees, type TreePlacement } from './treePlacement.js';
 import { TREES } from '@/config/components/trees.js';
 import { BUILDING_DIMENSIONS } from '@/config/components/buildings.js';
 import { FOOTPRINT } from '@/config/components/footprint.js';
-import { ISLAND_GEOMETRY } from '@/config/components/islandMesh.js';
+import { ISLAND_GEOMETRY } from '@/config/components/island.js';
 import { WORLD } from '@/config/world/world.js';
 import type { CityBbox, CityLayout } from '@/types';
